refactor: change tolerance from int -> float (#59)
Prior to this PR, width `tol`erance was expected to be of type `int`
and within the range of `[1, 100]` or "one to one hundred percent."
Now, `tol` is expected to be a `float` in the range `[0.01, 1.0]` which
also represents "one to one hundred percent."

We have done this for consistency purposes, i.e. [imgix-core-js](https://github.com/imgix/imgix-core-js/blob/2c9d97a7ab54c7a8fec378732bdcb855cef41907/src/imgix-core-js.js#L25) uses
floating point values. Since `imgix-core-js` was impl'd first, we defer
to this tried/tested interface. This way, this feature is standard
and consistent kit-wide.